% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ao_score.R
\name{ao_score}
\alias{ao_score}
\title{Corgi Quiz Assessment}
\usage{
ao_score(path, sheet_url, sheet_name, column_name, max_score, email)
}
\arguments{
\item{path}{path to csv file (string).}

\item{sheet_url}{Google spreadsheet link (string).}

\item{sheet_name}{Sheet name on the spreadsheet (string).}

\item{column_name}{The column that you want to update (string).}

\item{max_score}{Maximum score on the quiz (numeric).}

\item{email}{your email (string). for authentication purpose.}
}
\value{
"Quiz Updated Successfully" which mean the score on the spreadsheet is updated.
}
\description{
The corgi_score() function is used to assess the Algoritma students based on the quiz on corgi.re.
}
\examples{
ao_score(path = "data_input/QuizReporting_csv/ExploratoryDataAnalysisEDAQuiz-all_0.csv",
      sheet_url = "https://docs.google.com/spreadsheets/d/xxxxx/yyyyy",
      sheet_name = "Academy: Batch 9",
      column_name = "NN Quiz"
      max_score = 4,
      email = "david_at_algorit.ma")

}
