% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getColumnName.R
\name{getColumnName}
\alias{getColumnName}
\title{Get Column Name from Spreadsheet}
\usage{
getColumnName(word_length)
}
\arguments{
\item{word_length}{int}
}
\value{
character
}
\description{
Get Column Name from Spreadsheet
}
\examples{
getColumnName(21)
}
