% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedagogy_feedback.R
\name{pedagogy_feedback}
\alias{pedagogy_feedback}
\title{Pedagogy Feedback}
\usage{
pedagogy_feedback(
  connection,
  slack_channel,
  slack_username,
  slack_token,
  workshop_id
)
}
\arguments{
\item{connection}{Pedagogy database connection}

\item{slack_channel}{Slack channel name}

\item{slack_username}{Slack username}

\item{slack_token}{Slack token}

\item{workshop_id}{Id of Algoritma workshop}
}
\value{
send the summary of workshop to selected slack channel
}
\description{
The pedagogy_feedback() function is used to generate feedback from the pedagogy website.
}
\examples{
pedagogy_feedback(connection = cn,
          slack_channel = "datascience",
          slack_username = "David",
          slack_token = "x123yz",
          workshop_id = 890)

}
