% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recap_attendence.R
\name{recap_attendence}
\alias{recap_attendence}
\title{Recapitulation of Student Attendance}
\usage{
recap_attendence(
  gform_sheet,
  absen_sheet,
  absen_name,
  zoom_sheet,
  zoom_name,
  email
)
}
\arguments{
\item{gform_sheet}{Google Form spreadsheet link (string)}

\item{absen_sheet}{"Rekap Absensi Academy" spreadsheet link (string)}

\item{absen_name}{Sheet name on "Rekap Absensi Academy" (string)}

\item{zoom_sheet}{"Rekap Absensi Zoom" spreadsheet link (string)}

\item{zoom_name}{Sheet name on "Rekap Absensi Zoom" (string)}

\item{email}{Your algoritma email}
}
\value{
Student whose username was not found in the class
}
\description{
The recap_attendence() function is used to summarizes the attendance list from zoom and gform sheet
}
\examples{

recap_attendence(gform_sheet =  "https://docs.google.com/spreadsheets/d/xxx/yyy",
absen_sheet = "https://docs.google.com/spreadsheets/d/aaa/bbb",
absen_name =  "Newton DA Night",
zoom_sheet = "https://docs.google.com/spreadsheets/d/zzz/www",
zoom_name = "P4DS Day 2", email = "david_at_algorit.ma")

}
