% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_timeline.R
\name{update_timeline}
\alias{update_timeline}
\title{Update Timeline}
\usage{
update_timeline(schedule_row = 0, mentoring_row = 0, email)
}
\arguments{
\item{schedule_row}{row number of "schedule workshop" sheet (integer). The default is 0 which means the timeline sheet will not be updated from this sheet}

\item{mentoring_row}{row number of "piket report" sheet (integer). The default is 0 which means the timeline sheet will not be updated from this sheet}

\item{email}{your algoritma email (string)}
}
\value{
"Timeline_dev" sheet on "piket report" spreadsheet updated
}
\description{
The update_timeline() function is used to update timeline_dev sheet on piket report spreadsheet.
The source of data is from schedule workshop sheet and piket report sheet.
}
\examples{
update_timeline(schedule_row = 944,
mentoring_row = 440,
email = "David_at_algorit.ma")
}
