% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom_attendance.R
\name{zoom_attendance}
\alias{zoom_attendance}
\title{Zoom Attendance List}
\usage{
zoom_attendance(folder_name, sheet_url, sheet_name, email)
}
\arguments{
\item{folder_name}{The name of folder where attendance lists are stored}

\item{sheet_url}{Google spreadsheet link (string).}

\item{sheet_name}{Sheet name on the spreadsheet (string).}

\item{email}{your email (string). for authentication purpose.}
}
\value{
"Attendance Updated Successfully" which mean the score on the spreadsheet is updated.
}
\description{
The zoom_attendance() function is used to summarizes the attendance of students in classes conducted via zoom
}
\examples{
zoom_attendance(folder_name = "midas RM Day 3",
email = "david_at_algorit.ma",
sheet_url = "https://docs.google.com/spreadsheets/d/xx/yy",
sheet_name = "RM Day 3")

}
